dh_nginx: support auto generating module config files
This patch adds support of auto generating module config files for
modules in the naming pattern of libnginx-mod-* but with missing
<module_name>.nginx file. The module load file and its loading priority
is inferred from the package name in this case.

Using this feature, the repeated files d/libnginx-mod-*.nginx and
d/libnginx-mod.conf/ can be removed, and the packaging for out-of-tree
modules can be simplified.