nginx-dev: simplify dependencies
nginx-dev depends on ${S:Build-Depends} before this commit, to include
all the build-depdencies of nginx source packages. However, this also
includes build dependency for those in-tree 3rd party modules, which are
unnecessary. This commit removes them and explicitly list those packages
needed to build a general module. The architecture of nginx-dev can thus
be changed to all instead of any.

When all 3rd party modules are removed, the dependency can be changed
back to ${S:Build-Depends} to make the config less redundant.