d/nginx-*.postinst: Use invoke-rc.d instead of pidof and ad-hoc tests
Checking if nginx is running using and a mixture of file tests and
`pidof` is fragile. Debian policy 9.3.3 states that `invoke-rc.d` should
be used instead to perform these checks.

In addition, the use of `pidof` creates an implicit depency on
`sysvinit-utils`. `sysvinit-utils` is currently essential, but that
may change in the future.

Patch originally contributed by @ah: https://paste.debian.net/1254794/